+++ src/fiber/config.rs
@@ -40,7 +40,7 @@ pub const DEFAULT_TLC_MIN_VALUE: u128 = 0;
 pub const DEFAULT_TLC_MAX_VALUE: u128 = 0;
 
 /// The fee for forwarding peer tlcs. Proportional to the amount of the forwarded tlc. The unit is millionths of the amount. 1000 means 0.1%.
-pub const DEFAULT_TLC_FEE_PROPORTIONAL_MILLIONTHS: u32 = 1000;
+pub const DEFAULT_TLC_FEE_PROPORTIONAL_MILLIONTHS: u128 = 1000;
 
 // See comment in `LdkConfig` for why do we need to specify both name and long,
 // and prefix them with `ckb-`/`CKB_`.
@@ -129,7 +129,7 @@ pub struct FiberConfig {
         env,
         help = "The fee for forwarding peer tlcs. Proportional to the amount of the forwarded tlc. The unit is millionths of the amount. [default: 1000 (0.1%)]"
     )]
-    pub tlc_fee_proportional_millionths: Option<u32>,
+    pub tlc_fee_proportional_millionths: Option<u128>,
 }
 
 #[derive(PartialEq, Copy, Clone, Default)]
@@ -245,7 +245,7 @@ impl FiberConfig {
         self.tlc_max_value.unwrap_or(DEFAULT_TLC_MAX_VALUE)
     }
 
-    pub fn tlc_fee_proportional_millionths(&self) -> u32 {
+    pub fn tlc_fee_proportional_millionths(&self) -> u128 {
         self.tlc_fee_proportional_millionths
             .unwrap_or(DEFAULT_TLC_FEE_PROPORTIONAL_MILLIONTHS)
     }
